refactor(shortcuts) Pull all the global shortcuts into one place. - Fixes #274. - Extracted the logic from `editor-interactions.ts` and `canvas.ts`   pulling that into `global-shortcuts.ts`. In the process refactoring   out the inspector tab handling check. - Separated out browser shortcut prevention so that those can still   be applied even if the shortcuts don't match them.
