+++ editor/src/components/canvas/canvas.ts
@@ -330,363 +330,6 @@ const Canvas = {
       return null
     }
   },
-  handleKeyDown(
-    event: KeyboardEvent,
-    editor: EditorState,
-    derived: DerivedState,
-    dispatch: EditorDispatch,
-    insertImageFn: () => void,
-  ): Array<EditorAction> {
-    const key = Keyboard.keyCharacterForCode(event.keyCode)
-
-    const modifiers = modifiersForEvent(event)
-
-    const cmd = strictCheckModifiers(modifiers, ['cmd'])
-    const shift = strictCheckModifiers(modifiers, ['shift'])
-    const altCmd = strictCheckModifiers(modifiers, ['alt', 'cmd'])
-    const shiftCmd = strictCheckModifiers(modifiers, ['shift', 'cmd'])
-    const ctrl = strictCheckModifiers(modifiers, ['ctrl'])
-    const shiftCtrl = strictCheckModifiers(modifiers, ['shift', 'ctrl'])
-    const noModifier = modifiers.length === 0
-
-    const modeType = editor.mode.type
-    const adjustment = frameAdjustment(looseCheckModifier(modifiers, 'shift'))
-
-    switch (key) {
-      case 'backspace':
-      case 'delete':
-        if (editor.mode.type === 'select') {
-          if (cmd) {
-            return editor.selectedViews.map((target) => EditorActions.unwrapGroupOrView(target))
-          }
-        }
-        return [EditorActions.deleteSelected()]
-      case '0':
-        if (cmd) {
-          return [CanvasActions.zoom(1)]
-        } else {
-          return []
-        }
-      case 'plus':
-        if (altCmd) {
-          return [CanvasActions.zoomUI(true)]
-        } else if (cmd) {
-          return [CanvasActions.zoom(Utils.increaseScale(editor.canvas.scale))]
-        } else {
-          return []
-        }
-      case 'minus':
-        if (altCmd) {
-          return [CanvasActions.zoomUI(false)]
-        } else if (cmd) {
-          return [CanvasActions.zoom(Utils.decreaseScale(editor.canvas.scale))]
-        } else {
-          return []
-        }
-      case 'enter':
-        if (noModifier && modeType === 'select') {
-          const textTarget = getTextEditorTarget(editor)
-          if (textTarget != null) {
-            return [EditorActions.openTextEditor(textTarget, null)]
-          } else {
-            const childToSelect = Canvas.getFirstChild(
-              editor.selectedViews,
-              editor.jsxMetadataKILLME,
-            )
-            if (childToSelect == null) {
-              return []
-            } else {
-              return [EditorActions.selectComponents([childToSelect], false)]
-            }
-          }
-        } else if (shift && modeType === 'select') {
-          return jumpToParentActions(editor.selectedViews)
-        } else if (cmd && modeType === 'select') {
-          return [EditorActions.wrapInView(editor.selectedViews)]
-        } else {
-          return []
-        }
-      case 'esc':
-        if (noModifier) {
-          if (modeType === 'insert') {
-            return [
-              EditorActions.switchEditorMode(EditorModes.selectMode()),
-              CanvasActions.clearDragState(false),
-              EditorActions.clearHighlightedViews(),
-              EditorActions.setLeftMenuTab(LeftMenuTab.UINavigate),
-            ]
-          } else if (editor.canvas.dragState != null && editor.canvas.dragState.start != null) {
-            return [CanvasActions.clearDragState(false)]
-          } else if (modeType === 'select') {
-            return jumpToParentActions(editor.selectedViews)
-          }
-        }
-        return []
-      case 'space':
-        if (noModifier && modeType === 'select') {
-          if (CanvasMousePositionRaw == null) {
-            return [EditorActions.clearSelection()]
-          }
-          const targetStack = Canvas.getAllTargetsAtPoint(
-            editor,
-            CanvasMousePositionRaw,
-            [TargetSearchType.All],
-            true,
-            'strict', // _IF_ we want to enable loose targeting for selection, it means we also need to change component-area-control
-          )
-          const nextTarget = Canvas.getNextTarget(editor.selectedViews, targetStack)
-          if (targetStack.length === 0 || nextTarget === null) {
-            return [EditorActions.clearSelection()]
-          } else {
-            return [EditorActions.selectComponents([nextTarget], false)]
-          }
-        } else {
-          return []
-        }
-      case 'tab':
-        if ((shift || noModifier) && modeType === 'select') {
-          const forwards = !shift
-          const tabbedTo = Canvas.jumpToSibling(
-            editor.selectedViews,
-            editor.jsxMetadataKILLME,
-            forwards,
-          )
-          if (tabbedTo == null) {
-            return []
-          } else {
-            return [EditorActions.selectComponents([tabbedTo], false)]
-          }
-        } else {
-          return []
-        }
-      case 'up':
-      case 'down':
-        if (altCmd) {
-          return key === 'up'
-            ? [EditorActions.moveSelectedForward()]
-            : [EditorActions.moveSelectedBackward()]
-        } else {
-          const isResizing = looseCheckModifier(modifiers, 'cmd')
-          const verticalAdjustmentActions = adjustAllSelectedFrames(
-            editor,
-            dispatch,
-            false,
-            isResizing,
-            key === 'up' ? -1 : 1,
-            'vertical',
-            adjustment,
-          )
-          return [EditorActions.transientActions(verticalAdjustmentActions)]
-        }
-      case 'left':
-      case 'right':
-        const isResizing = looseCheckModifier(modifiers, 'cmd')
-        const horizontalAdjustmentActions = adjustAllSelectedFrames(
-          editor,
-          dispatch,
-          false,
-          isResizing,
-          key === 'left' ? -1 : 1,
-          'horizontal',
-          adjustment,
-        )
-        return [EditorActions.transientActions(horizontalAdjustmentActions)]
-      case 'a':
-        if (cmd) {
-          return [EditorActions.selectAllSiblings()]
-        } else {
-          return []
-        }
-      case 'b':
-        if (noModifier && editor.selectedViews.length > 0) {
-          return TP.filterScenes(editor.selectedViews).map((target) =>
-            EditorActions.toggleProperty(
-              target,
-              toggleStylePropPath(PP.create(['style', 'border']), toggleBorder),
-            ),
-          )
-        } else {
-          return []
-        }
-      case 'c':
-        if (cmd) {
-          return [EditorActions.copySelectionToClipboard()]
-        } else if (editor.selectedViews.length > 0) {
-          // TODO Toggle Clipping
-          return []
-        } else {
-          return []
-        }
-      case 'd':
-        if (cmd && editor.selectedViews.length > 0) {
-          return [EditorActions.duplicateSelected()]
-        } else if (cmd) {
-          return []
-        }
-        return []
-      case 'f':
-        if (noModifier && editor.selectedViews.length > 0) {
-          return TP.filterScenes(editor.selectedViews).map((target) =>
-            EditorActions.toggleProperty(target, toggleStylePropPaths(toggleBackgroundLayers)),
-          )
-        } else {
-          return []
-        }
-      case 'g':
-        if (editor.selectedViews.length > 0) {
-          if (shiftCmd) {
-            return editor.selectedViews.map((target) => EditorActions.unwrapGroupOrView(target))
-          } else if (cmd) {
-            return [EditorActions.wrapInGroup(editor.selectedViews)]
-          }
-        }
-        return []
-      case 'h':
-        if (shiftCmd && editor.selectedViews.length > 0) {
-          return [EditorActions.toggleHidden()]
-        } else {
-          return []
-        }
-      case 'i':
-        if (noModifier && (modeType === 'select' || modeType === 'insert')) {
-          insertImageFn()
-          return []
-        } else {
-          return []
-        }
-      case 'p':
-        if (shiftCmd && editor.selectedViews.length > 0) {
-          return [] // TODO Nodegraph insertConnectedNodeActions('GetProperty', editor.selectedViews[0])
-        } else if (noModifier && modeType === 'select') {
-          return [] // sorry but path is too broken for now [enableInsertMode({package: 'svg', template: 'path'}, defaultPathCommands())]
-        } else {
-          return []
-        }
-      case 'r':
-        if (cmd) {
-          const exitInsertModeActions = [
-            EditorActions.switchEditorMode(EditorModes.selectMode()),
-            CanvasActions.clearDragState(false),
-            EditorActions.clearHighlightedViews(),
-            EditorActions.setLeftMenuTab(LeftMenuTab.UINavigate),
-          ]
-          if (editor.selectedViews.length === 1) {
-            const target = editor.selectedViews[0]
-            return [EditorActions.setNavigatorRenamingTarget(target), ...exitInsertModeActions]
-          } else {
-            return exitInsertModeActions
-          }
-        }
-        if (noModifier && (modeType === 'select' || modeType === 'insert')) {
-          const utopiaComponents = getOpenUtopiaJSXComponentsFromState(editor)
-          const newUID = generateUidWithExistingComponents(utopiaComponents)
-          return [
-            EditorActions.enableInsertModeForJSXElement(
-              defaultRectangleElement(newUID),
-              newUID,
-              { 'utopia-api': importDetails(null, [importAlias('Rectangle')], null) },
-              null,
-            ),
-          ]
-        }
-        return []
-      case 'e':
-        if (noModifier && (modeType === 'select' || modeType === 'insert')) {
-          const utopiaComponents = getOpenUtopiaJSXComponentsFromState(editor)
-          const newUID = generateUidWithExistingComponents(utopiaComponents)
-          return [
-            EditorActions.enableInsertModeForJSXElement(
-              defaultEllipseElement(newUID),
-              newUID,
-              { 'utopia-api': importDetails(null, [importAlias('Ellipse')], null) },
-              null,
-            ),
-          ]
-        }
-
-        return []
-      case 's':
-        if (noModifier && modeType === 'select' && editor.selectedViews.length > 0) {
-          return TP.filterScenes(editor.selectedViews).map((target) =>
-            EditorActions.toggleProperty(
-              target,
-              toggleStylePropPath(PP.create(['style', 'boxShadow']), toggleShadow),
-            ),
-          )
-        } else if (shiftCmd && editor.selectedViews.length > 0) {
-          return [] // TODO Nodegraph insertConnectedNodeActions('SetProperty', editor.selectedViews[0])
-        } else {
-          return []
-        }
-      case 't':
-        if (noModifier && (modeType === 'select' || modeType === 'insert')) {
-          const utopiaComponents = getOpenUtopiaJSXComponentsFromState(editor)
-          const newUID = generateUidWithExistingComponents(utopiaComponents)
-          return [
-            EditorActions.enableInsertModeForJSXElement(
-              defaultTextElement(newUID),
-              newUID,
-              { 'utopia-api': importDetails(null, [importAlias('Text')], null) },
-              null,
-            ),
-          ]
-        } else if (shiftCmd && editor.selectedViews.length > 0) {
-          return [] // TODO Nodegraph insertConnectedNodeActions('TouchInteraction', editor.selectedViews[0])
-        } else {
-          return []
-        }
-      case 'v':
-        if (altCmd) {
-          if (editor.selectedViews.length > 0) {
-            // TODO Paste Style
-            return []
-          } else {
-            return []
-          }
-        } else if (noModifier && (modeType === 'select' || modeType === 'insert')) {
-          const utopiaComponents = getOpenUtopiaJSXComponentsFromState(editor)
-          const newUID = generateUidWithExistingComponents(utopiaComponents)
-          return [
-            EditorActions.enableInsertModeForJSXElement(
-              defaultViewElement(newUID),
-              newUID,
-              { 'utopia-api': importDetails(null, [importAlias('View')], null) },
-              null,
-            ),
-          ]
-        } else {
-          return []
-        }
-      case 'x':
-        if (cmd) {
-          return [EditorActions.copySelectionToClipboard(), EditorActions.deleteSelected()]
-        } else {
-          return []
-        }
-      case 'z':
-        return [EditorActions.setHighlightsEnabled(false), EditorActions.clearHighlightedViews()]
-      case '[':
-        if (cmd) {
-          event.preventDefault()
-          return [EditorActions.moveSelectedBackward()]
-        } else if (altCmd) {
-          return [EditorActions.moveSelectedToBack()]
-        } else {
-          return []
-        }
-      case ']':
-        if (cmd) {
-          event.preventDefault()
-          return [EditorActions.moveSelectedForward()]
-        } else if (altCmd) {
-          return [EditorActions.moveSelectedToFront()]
-        } else {
-          return []
-        }
-      default:
-        return []
-    }
-  },
   handleKeyUp(key: KeyCharacter, editor: EditorState, derived: DerivedState): Array<EditorAction> {
     switch (key) {
       case 'z':
@@ -695,42 +338,6 @@ const Canvas = {
         return []
     }
   },
-}
-
-function jumpToParentActions(selectedViews: Array<TemplatePath>): Array<EditorAction> {
-  const jumpResult = Canvas.jumpToParent(selectedViews)
-  switch (jumpResult) {
-    case null:
-      return []
-    case 'CLEAR':
-      return [EditorActions.clearSelection()]
-    default:
-      return [EditorActions.selectComponents([jumpResult], false)]
-  }
-}
-
-function getTextEditorTarget(editor: EditorState): InstancePath | null {
-  if (editor.canvas.dragState != null || editor.selectedViews.length !== 1) {
-    return null
-  } else {
-    const target = editor.selectedViews[0]
-    if (TP.isScenePath(target)) {
-      return null
-    }
-
-    const components = getOpenUtopiaJSXComponentsFromState(editor)
-    const imports = getOpenImportsFromState(editor)
-    const element = findElementAtPath(target, components, editor.jsxMetadataKILLME)
-    if (element != null && isUtopiaAPITextElement(element, imports)) {
-      return target
-    } else {
-      return null
-    }
-  }
-}
-
-function frameAdjustment(shiftPressed: boolean): 10 | 1 {
-  return shiftPressed ? 10 : 1
 }
 
 export default Canvas
